requirements.txt
PyYAML==6.0
joblib==1.2.0
pandas==1.4.3
scikit-learn==1.1.2